avoid while(true) loop in FoldStreamAsync
